Move restify handlers into routes/.
